% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R
\name{hlsyntax}
\alias{hlsyntax}
\title{Highlight code syntax}
\usage{
hlsyntax(x, lexer)
}
\arguments{
\item{x}{character model specification.}
}
\description{
Uses Python Pygements to highlight code.
}

