% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{pygmentr_styles}
\alias{pygmentr_styles}
\title{List all styles (including plugins) available to Pygments}
\usage{
pygmentr_styles()
}
\description{
List all styles (including plugins) available to Pygments
}

